Let model.loss = the sum of the losss of previous tuples

+++ src/include/utils/tuplesort.h
@@ -33,7 +33,7 @@ typedef struct Tuplesortstate Tuplesortstate;
 
 // Lijie: add begin
 typedef struct {
-	int loss;
+	double loss;
 	int p1;
 	int p2;
 } Model;
